Multi-Task_LLM_API/requirements.txt
Flask
python-dotenv
Flask-Limiter
Flask-RESTX
tenacity

waitress
langchain
langchain-google-genai
langchain-community